requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
supabase==2.0.2
asyncpg==0.29.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
pillow==10.1.0
aiofiles==23.2.1
httpx>=0.24.0,<0.25.0